Cleanup: move method generation into class, mostly
Move as much as possible of the metaprogramming/method generation
into the affected classes. The class object doesn't exist until
the end of the class definition, which is where we previously had
the code to add more methods. Actually, we can put that code inside
the class as a classmethod, but we can't invoke it until after the
class definition (because we need to pass the classmethod a
reference to the class, and that's the first point that exists).