Fix distinction between Device and OutputDevice.
Device goes away (well, is replaced by a BaseDevice) that nothing uses
yet; all existing use of Device uses OutputDevice instead, since that's
what it wants.

BaseDevice will later be used as a common superclass for outputs
(OutputDevice class) and keypads (ControlDevice class); the class name
"Device" is not good for this since Lutron uses the word "device" for
what I want to call a ControlDevice.