Add support for "controls" (keypads).
- Distinguish between "output" devices and "control" devices more carefully
- Add support for "devices" to the ra_layout layer, as a collection of
  buttons and leds
- Send requests and parse responses for button and led status in ra_repeater
- Redo ra_repeater's cache so it's a true cache, and reads block if the value
  isn't there yet, so we don't actually depend on pre-populating it
- Revisit class hierarchy in ra_house, moving the boundary between
  LutronDevice and OutputDevice once again (now that ControlDevice has a
  nontrivial implementation, it turns out it can share more behavior with
  OutputDevice via the base LutronDevice class)
- Flesh out ra_house's ControlDevice and KeypadDevice classes, actually
  implementing a keypad object with a list of buttons that can query and
  set button and led state
- Introduce a new "devclass" distinction visible to the UI so that it can
  operate on all devices, or only control or output devices
- Represent keypads, buttons and leds in UI (very basically)

The upper layers need more work; the UI for keypads is very rudimentary,
not all concepts make sense or have been reconciled (especially filtering
by state/type) for controls, and the relationship between what ra_house's
ideas of devclass and devtype, the UI (demo.py and templates) idea of
devclass and devtype, and the class hierarchy in ra_house is a complete
mess, with overlap and contradiction. And the templates named "output"
actually apply to all devices (controls and outputs both). To be cleaned
up in a future step.