Persistence improvements
Track device current state in persistence module so it can be used in
elapsed time calculations where it matters.

Store separate timestamps for device change events, one restricted by
process lifetime (so we know we've been watching and didn't miss any
events), and one not (so it may be more informative, but may also have
missed some events). In the case we've been running long enough to see
the device change state, these will be the same; otherwise, we end up
with our process start time and the previous database timestamp entry
as lower and upper bounds on that last change event, and in different
contexts, each of these is useful.

Database schema was rejiggered a little; ra_house object initialization
and the signatures and plumbing for on_user_change were rejiggered
significantly, in order to save device state at the time we first hear
it during the initial cache refresh.

And some serious bugfixes, exposed and/or found while doing the above,
but which were already problems before:
- make persistence module thread-safe in its use of sqlite3 (by taking
  a lock around all public methods, before calling into sqlite3)
- on->on transitions for dimmers (say between 50% and 80%) still have
  an end state of on. That is, we have to look at the old state, and
  not assume it's just the opposite of the new state.