Minor UI cleanup.
Get rid of duplicated activate_control_hack method, in favor of a
less verbose hack.

Learned about difference between form/args/values attributes of
request object. Used this so that get_control can redirect to
activate_control if args are specified, with a minimum of code
duplication.

Fixed a preexisting buglet in set_outputs_to_state introduced in
one of the previous renamings.