Don't try to checkpoint area states/changes.
Areas are in the devicemap but aren't really devices, and aren't
in the other tables. When the checkpointer grabs device IDs from
the devicemap, avoid the areas, which avoids a spurious log message
about checkpointing a device in unknown state, and reduces DB
clutter.

This may argue further for area IDs living in their own table
instead of the devicemap.