Bug fix: don't clobber last_seen_event on restart.
The original point of last_seen_event_ts is that it survives across
the process lifetime. I broke that when I changed init_device_state's
SQL implementation from "update" to "insert or replace".

What I really want is "upsert", which will update an existing row
(preserving fields I don't want to change) when it exists, and create
a new one (with null values for the fields I don't specify) when it
doesn't.

Credit to http://stackoverflow.com/questions/418898/sqlite-upsert-not-insert-or-replace
for this technique.