Use device's own implementation of order_states().
Instead of using the jinja filter registered at the toplevel,
hardwired to a specific implementation, and trying to pass the
class along so it has some flexibility, just call through the
device directly.

This is easy here where we always have a device; it's harder for
aggregates like areas, where we have a bunch of devices (which
may not have the same type and thus idea of order_states); this
needs fixing later.

For now, add a default do-nothing order_states to SgDevice.