Disentangle class attributes and instance attributes.
I've been abusing class attributes as default values for instance
attributes, because I liked the effect (self-documenting code at
the top of each class that's both machine readable and human
readable) and because it seemed to work. However, it reflected
confusion on my part and could cause confusion for other readers
and/or bugs if shared mutable values were changed when they weren't
intended to be shared.

(In practice, many of these are not changed after object creation
and do fine as class attributes, and even the ones that are changed
are changed by assignment, meaning the object will get an instance
attribute with the same name shadowing the class one. That's why
this seemed to work fine. In a few cases, the only thing that saved
me was that I only created one instance of the offending class. In
any case, these attributes logically belong to the instance and not
the class, so let's avoid the abuse.)

I learned a lot about Python attribute lookup order while unscrewing
this.