Introduce "synther" plugin which can coordinate events across other devices.
Right now it provides a bridge between Lutron and DSC devices, where a
Lutron output (programmed into the Lutron system but not activated, so
it has an integration id but no hardware) can map to a DSC zone sensor
(reading the state of the real output) and a DSC PGM output (to tell the
real output to change state).

Changes:
- config: added order mechanism, synther module config
- events: added publish/subscribe by individual device
- sg_house and gateways: populate house.gateways live, as we load gateways,
  and added (very hacky) order mechanism, so later gateways can refer to
  earlier ones
- ra_gateway: add optional argument to OutputDevice.be_on so you can call
  it as be_on(False) to turn off the device.
- dsc_gateway: added scoped/namespaced format for gateway_devid so we can
  tell the difference between partition id and zone id; added method to
  send user-command requests to panel via Envisalink
- synther: added new gateway plugin that bridges Lutron and DSC gateways
  as described above

What's here is hacky, unfinished and needs cleanup but it's enough to get
buttons on my Lutron keypads controlling old 3-way lighting wired into
relays controlled and monitored by the DSC panel.