Prevent infinite device-change loop when a device (not controlled by Lutron, but bound to a Lutron button with synther) changes outside the Lutron button control, synther tells the Lutron button it changed, and then synther sees the Lutron repeater say it changed again.
Clean up the ignore_time mechanism, and reuse it for this.

This might not be the right way of dealing with this -- it seems
there is another bug lurking, because I'd expect that the message
from the Lutron repeater saying it changed should carry the state
it changed to, which is already the current state, and we shouldn't
respond further. But good enough for now.