Simplify and cleanup.
Pull the call to send the global status request to the panel out
of DscPanelCache and into DscPanelServer.connect(). Then DscPanelCache
doesn't need a reference to the panel server at all. Also, split
cache mark-stale from init; then the panel server can create it
once and reuse it later.

This is cleaner, and is also in preparation for automatic reconnect.