Simplify connect method, reorder other methods, and use a separate sender thread (copied from DSC panel implementation).
When I originally wrote this code, it was in a request-response
mindset, and only over time did it become clear that the requests and
responses should be wholly asynchronous. By the time I wrote the DSC
panel implementation I'd acquired that mindset, and now it's time to
bring it back to the older Lutron repeater implementation.

This gets rid of the whole state table where I map repeater responses
to some idea of state, and only send commands when the repeater is in
the ready state. Now we just write when we want, and listen to whatever
it says.

Connect gets simpler; we connect the socket, authenticate with the
simple blocking socket read/write calls, then (assuming it succeeds;
no good error handling for authentication failure yet) put the socket
into nonblocking mode and start the reader/writer threads and let them
take it from there.

This is all in service of a larger goal which is automatic reconnect.