Automatic reconnect support for stateful gateway connections.
Lutron and DSC plugins both have to deal with the fact that they
establish a stateful and (hopefully) long-lived connection to the
gateway hardware; if this goes away (gateway device reboots, network
interruption, etc; especially for Lutron where the repeater will
tolerate one connection per credential and boot off any existing
client if someone else connects with the same credential) so that
the socket is unavailable, they need to reestablish the connection.
(Vera doesn't have to deal with this, since its protocol is stateless,
using a new HTTP connection for each request.)

This change introduces:
- new SgWatchdog class in connections.py, which uses select() to monitor
  all sockets registered with it, and invoke a cleanup routine if
  select() reports they're in an errored state. Actually it's worse
  than this; for the case I care about, closed socket, select() throws
  a socket.error exception that doesn't say which socket caused it,
  so then I reprobe all the sockets one at a time.
- StargateHouse owns and initializes the watchdog.
- Lutron and DSC gateway plugins register their sockets with the
  watchdog, and had the signature of their connect() routines rejiggered
  a bit so they could take advantage of it.

There are some ugly hacks and shortcuts taken here; notably in the
watchdog thread:
- there must be a better way to find out which socket is responsible
  for the socket.error exception, but I couldn't find one.
- cleanup process needs a way to tell the sender and listener threads
  to exit; listener blocks on socket read so that's accomplished by
  making sure the socket is closed (in practice, the listener will be
  the one to notice the problem anyway); sender blocks on Queue.get()
  so we have a nasty hack to force that to return.
- watchdog class has no locking around the notify-list management.
- watchdog class also has no notification around notify-list add,
  so the watcher thread has to wake up often to check for changes.

The last two of these should be easy enough to fix; the first two
less so.