Create smtplib.SMTP object when we want to use it, not up front.
Previously I thought it was a good idea to keep the SMTP object
around forever -- create one at process start, and reuse it for
any emailing we later do.

It turns out that it connects at init time, the way I was calling
it; it's probably not a good idea to connect at init time, ignore
it for days/weeks/months, then attempt to send.

So let's create the SMTP helper object whenever we need to use it.