Report exceptions via reporter mechanism.
Teach sg_signal to interpose on logger.exception (by adding its own
logger handler which looks for report.exc_info) and now exceptions
are one of the conditions it lets clients register for.

Teach reports to use that to listen for exceptions that got logged
(which should be all unhandled exceptions), and send the exception
info via the notify mechanism.