Maintain reference count, not just bool, for refreshing devices.
Multiple things could ask for a refresh before an outstanding
refresh finishes, and we don't want the first answer to come back
to clear the refreshing bit so the further responses are treated
as real events without the "synthetic" flag.

This isn't perfect (if a real person presses a button while the
thundering refresh herd is still chugging along, their press will
be treated as synthetic) but it's better than the current behavior
(and even in that case, the last would-be-refresh response will be
treated as not synthetic, so a real event will eventually be
generated).

Fixes #11, partially.