Support DSC motion sensors (low level).
Add support for different DSC device types, in the config file and
dsc_gateway config parser. Currently the only device types are
"closure" (already supported) and "motion" (a minor variant that
reports its status differently).

The config file format was augmented to allow DSC zones to label
their type; the previous format (just a string name) is interpreted
as shorthand for a closure zone.

This is low-level support to make Stargate aware of the "motion"
devtype for queries and status display; to be truly useful it should
report itself as "occupied" if the last change was within the reporting
cooldown time (3 minutes?), and/or the UI should prioritize showing its
age instead of current occupied/vacant status.